requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
Django==4.2.25
djangorestframework==3.16.1
mysqlclient==2.1.1
pillow==12.0.0
reportlab==4.4.4
sqlparse==0.5.3
typing_extensions==4.15.0
